Mover los errores de dominio fuera de la capa de servicios
AssignmentNoEncontradoError, AssignmentNoDisponibleError y
GrupoNoEncontradoError vivían en assignmentAuthorization.ts (capa de
servicios), pero los repositorios los importaban directamente —una
inversión de capas: la infraestructura terminaba dependiendo hacia
arriba de la lógica de aplicación. Se mudan junto a las entidades que
describen (Assignment, Grupo) y se reexportan desde
assignmentAuthorization.ts para no romper los imports existentes.
AccesoAssignmentProhibidoError queda donde está: es un veredicto de
autorización, no un hecho sobre el estado de una entidad.

De paso, NOMBRES_ESTADO_ASSIGNMENT pasa a ser de solo lectura para que
ningún consumidor pueda mutar la constante compartida.
